# LibGfx/JPEGXL: Avoid allocations in the hot path