This was obviously very naive, pre-allocating the array makes it a lot faster (~1.56 times faster, on the JPEG XL logo image [1]).

Before:
Benchmark 1: image JPEGXL/logo.jxl --no-output
  Time (mean ± σ):     922.2 ms ±  18.4 ms
  Range (min … max):   891.9 ms … 945.4 ms    10 runs

After:
Benchmark 1: image JPEGXL/logo.jxl --no-output
  Time (mean ± σ):     585.4 ms ±  10.0 ms
  Range (min … max):   570.5 ms … 597.9 ms    10 runs

[1] https://jpegxl.info/old/logo.jxl